pyproject.toml
[tool.poetry]
name = "fave-text-normalize"
version = "0.1.0"
description = "Text normalize a FAVE-style transcription"
authors = ["Josef Fruehwald <JoFrhwld@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pympi-ling = "^1.70.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
